:recycle: Extract interface from FakeImageUploader
