+++ src/main/java/com/mercadolivre/apimlv2/shared/imageupload/FakeImageUploader.java
@@ -1,4 +1,4 @@
-package com.mercadolivre.apimlv2.usecases.addimagetoproduct;
+package com.mercadolivre.apimlv2.shared.imageupload;
 
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -10,13 +10,9 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
-public class ImageUploader {
+public class FakeImageUploader implements ImageUploader {
 
-    /**
-     * Upload a image to a bucket and return its url.
-     * @param files Images to be uploaded
-     * @return Set containing urls of uploaded files
-     */
+    @Override
     public Set<String> upload(List<MultipartFile> files) {
         return files.stream()
                     .map(this::getUrl)
